# Speed up bcf genotype count